+++ tools/src/bin/ec4rs-parse.rs
@@ -27,7 +27,6 @@ fn print_config(path: &std::path::Path, filename: Option<&PathBuf>, legacy_fallb
                 props.use_fallbacks();
             }
             for (key, value) in props.iter() {
-                let value = value.into_str();
                 if ec4rs::property::STANDARD_KEYS.contains(&key) {
                     println!("{}={}", key, value.to_lowercase())
                 } else {